rename controller name
